Abstract Items with GlowMasks
- Create base class for Items with glowmasks.
- Use said base class.

PS. Projectiles are quite different and there is one that would use it. so no class for them.
