+++ MiscGlowMasks.cs
@@ -1,5 +1,6 @@
 ﻿
 using System;
+using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using Terraria;
 using Terraria.GameContent;
@@ -35,6 +36,38 @@ namespace GoldensMisc
 				RodofWarping = null;
 				Loaded = false;
 			}
+		}
+
+	}
+
+	public abstract class GlowingModItem : ModItem
+    {
+		private Texture2D _glowMask;
+
+		public Texture2D GlowMask 
+		{
+			get { return _glowMask; }
+			set { _glowMask = value; }
+		}
+
+		public override void PostDrawInInventory(SpriteBatch spriteBatch, Vector2 position, Rectangle frame, Color drawColor, Color itemColor, Vector2 origin, float scale)
+		{
+			Texture2D glowTexture = _glowMask;
+			Color color = Color.White;
+			color.A = 255;
+			spriteBatch.Draw(glowTexture, position, frame, color, 0f, origin, scale, SpriteEffects.None, 0f);
+		}
+
+		public override void PostDrawInWorld(SpriteBatch spriteBatch, Color lightColor, Color alphaColor, float rotation, float scale, int whoAmI)
+		{
+			Texture2D glowTexture = _glowMask;
+			Color color = Color.White;
+			color.A = 255;
+			Rectangle glowTextureSize = new(0, 0, glowTexture.Width, glowTexture.Height);
+			Vector2 glowTextureOrigin = glowTexture.Size() * .5f;
+			Vector2 glowOffset = new((float)(Item.width / 2) - glowTextureOrigin.X, (float)(Item.height - glowTexture.Height));
+			Vector2 glowPosition = Item.position - Main.screenPosition + glowTextureOrigin + glowOffset;
+			spriteBatch.Draw(glowTexture, glowPosition, glowTextureSize, color, rotation, glowTextureOrigin, scale, SpriteEffects.None, 0f);
 		}
 	}
 }
